Fix module imports and fix module small bugs
