+++ sshsysmon/inspectors/filemeta.py
@@ -1,5 +1,5 @@
-from inspector import Inspector
-from StringIO import StringIO
+from lib.plugins import Inspector
+from io import StringIO
 from fnmatch import fnmatch
 from lib.util import ByteSize, TimeSpan
 from datetime import datetime, timedelta
