+++ sshsysmon/inspectors/system.py
@@ -1,4 +1,4 @@
-from inspector import Inspector
+from lib.plugins import Inspector
 from lib.util import TimeSpan
 
 """
