+++ sshsysmon/lib/plugins/inspector.py
